+++ astropy/io/ascii/__init__.py
@@ -40,6 +40,7 @@ from .sextractor import SExtractor
 from .fixedwidth import (FixedWidth, FixedWidthNoHeader,
                          FixedWidthTwoLine, FixedWidthSplitter,
                          FixedWidthHeader, FixedWidthData)
+from .rst import RST
 from .ui import (set_guess, get_reader, read, get_writer, write, get_read_trace)
 
 from . import connect